fix(timr): changed behaviour of finished to only emit finish
Previous behaviour when a timer finished used this.stop() to clear the timers. This resulted in a
timer emitting stop AND finish when it had finished which is not the desirable behaviour.

A new "finished" status is also set when a timer has finished.

BREAKING CHANGE: When a timer has finished it no longer emits "onStop" and only emits "finish".

fix #21