feat(date): past dates no longer throws an error
When providing a date that's in the past (without a backupStartTime, or a backup that's also in the
past) the timer now no longer throws an error and immediatley finishes the timer. This allows for
better handling of non persistent timers.

BREAKING CHANGE: start() no longer throws an error when startTime = 0 with countdown = true.
dateToSeconds also no longer throws an error when startTime or backupStartTime are in the past but
now returns 0.