feat(date): using date-fns/parseISO to parse dates
Rather then the original custom date parsing, we're now using date-fns/parseISO to be fully
compatible with parsing ISO8601 dates.

BREAKING CHANGE: When passing an invalid date format, e.g. (0000-00-00). Previously an error would
be thrown when it couldn't be parsed with a message starting: 'The provided date is not in the right
format or is incorrect.'. Now it will throw an error with a message: 'Expected time to be in
(hh:mm:ss) format, instead got: 0000-00-00'. This is due to the fact that the parsing logic is
passed over to the timeToSeconds function when a parseISO returns an "Invalid Date".